requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pdfminer.six==20231228
python-docx==1.1.2
google-genai==1.3.0

